+++ src/commonMain/kotlin/IBoard.kt
@@ -1,6 +1,6 @@
 interface IBoard {
-    fun getPiece(x: Byte, y: Byte): Piece?
-    fun setPiece(x: Byte, y: Byte, piece: Piece?)
+    fun getPiece(x: Int, y: Int): Piece?
+    fun setPiece(x: Int, y: Int, piece: Piece?)
     fun getKing(pieceColor: PieceColor): Square?
     fun getSquaresByPieceColor(pieceColor: PieceColor): List<Square>
     fun isPathClear(move: PotentialMove): Boolean
