+++ src/commonMain/kotlin/Queen.kt
@@ -1,7 +1,7 @@
 class Queen(override val color: PieceColor) :
     Piece(PieceType.QUEEN, color, if (color == PieceColor.WHITE) 'Q' else 'q') {
     override fun canMove(move: PotentialMove): Boolean {
-        return move.deltaY == 0.toByte() || move.deltaX == 0.toByte() || move.distanceY == move.distanceX
+        return move.deltaY == 0 || move.deltaX == 0 || move.distanceY == move.distanceX
     }
 
     override val trajectories: List<MoveTrajectory> =
